Fixed VC2015 RC warnings.
-possible loss of data while conversions
-potentially uninitialized local variables
-declaration hides previous local declaration

+++ example/example2.c
@@ -40,8 +40,8 @@ int main()
 		printf("Could not open SVG image.\n");
 		goto error;
 	}
-	w = image->width;
-	h = image->height;
+	w = (int)image->width;
+	h = (int)image->height;
 
 	rast = nsvgCreateRasterizer();
 	if (rast == NULL) {
